pyproject.toml
[tool.poetry]
name = "wingz-exam"
version = "0.1.0"
description = ""
authors = ["Leonard Mark <leonardmarkdimayuga@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.1.5"
djangorestframework = "^3.15.2"
markdown = "^3.7"
django-filter = "^24.3"
django-stubs = "^5.1.2"
django-model-utils = "^5.0.0"
django-cors-headers = "^4.6.0"
python-dotenv = "^1.0.1"
factory-boy = "^3.3.1"


[tool.poetry.group.dev.dependencies]
flake8 = "^7.1.1"
flake8-pyproject = "^1.2.3"
black = "^24.10.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.black]
line-length = 120

[tool.flake8]
# Check that this is aligned with your other tools like Black
max-line-length = 120
exclude = [
    # No need to traverse our git directory
    ".git",
    # There's no value in checking cache directories
    "__pycache__"
]
# Use extend-ignore to add to already ignored checks which are anti-patterns like W503.
extend-ignore = [
    # PEP 8 recommends to treat : in slices as a binary operator with the lowest priority, and to leave an equal
    # amount of space on either side, except if a parameter is omitted (e.g. ham[1 + 1 :]).
    # This behaviour may raise E203 whitespace before ':' warnings in style guide enforcement tools like Flake8.
    # Since E203 is not PEP 8 compliant, we tell Flake8 to ignore this warning.
    # https://black.readthedocs.io/en/stable/the_black_code_style/current_style.html#slices    
    "E203"
]